pagar input cambiar cuenta

+++ src/main/java/ar/edu/unlam/tallerweb1/servicios/ServicioCuota.java
@@ -28,6 +28,6 @@ public interface ServicioCuota {
 
 	void pagarCuotaSeleccionada(Confirmpagocuota confirm);
 
-	boolean pagarporinput(Integer pago, Long idPrestamo);
+	boolean pagarporinput(double pago, Long idPrestamo);
 
 }
